backend/requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-cors>=4.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
cloudinary>=1.36.0
werkzeug>=3.0.0
gunicorn